# project/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Auditable Agent Runtime API
  version: 1.0.0
  description: Minimum MVP contract for a multi-tenant auditable Agent Runtime.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /auth/refresh:
    post:
      security: []
      operationId: refresh
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { description: Tokens refreshed, content: { application/json: { schema: { $ref: '#/components/schemas/AuthResponse' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /agents:
    get:
      operationId: listAgents
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/PageSize' }]
      responses:
        '200': { description: Agent page, content: { application/json: { schema: { $ref: '#/components/schemas/AgentPage' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createAgent
      parameters: [{ $ref: '#/components/parameters/IdempotencyKey' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentCreate' }
      responses:
        '201': { description: Agent created, content: { application/json: { schema: { $ref: '#/components/schemas/AgentEnvelope' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /agents/{agent_id}/publish:
    post:
      operationId: publishAgent
      parameters:
        - { $ref: '#/components/parameters/AgentId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '201': { description: Immutable version published, content: { application/json: { schema: { $ref: '#/components/schemas/AgentVersionEnvelope' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /agents/{agent_id}:
    get: { summary: Get agent, description: Returns one tenant-scoped agent., operationId: getAgent, parameters: [{ $ref: '#/components/parameters/AgentId' }], responses: { '200': { description: Agent returned }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
    patch: { summary: Update agent, description: Updates an agent draft., operationId: updateAgent, parameters: [{ $ref: '#/components/parameters/AgentId' }], requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Agent updated }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
    delete: { summary: Delete agent, description: Soft deletes an agent., operationId: deleteAgent, parameters: [{ $ref: '#/components/parameters/AgentId' }], responses: { '204': { description: Agent deleted }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
  /agents/{agent_id}/versions:
    get: { summary: List agent versions, description: Lists immutable published versions., operationId: listAgentVersions, parameters: [{ $ref: '#/components/parameters/AgentId' }], responses: { '200': { description: Versions returned }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
  /agents/{agent_id}/rollback:
    post: { summary: Roll back agent, description: Publishes a rollback as a new immutable version., operationId: rollbackAgent, parameters: [{ $ref: '#/components/parameters/AgentId' }], requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '201': { description: Rollback published }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
  parameters:
    AgentId: { name: agent_id, in: path, required: true, schema: { type: string, minLength: 10, maxLength: 100 } }
...
    Cursor: { name: cursor, in: query, schema: { type: string, maxLength: 512 } }
    PageSize: { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 16, maxLength: 100 } }
    IfMatch: { name: If-Match, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 100 } }
...
  responses:
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: Action is not allowed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Conflict: { description: Conflict or idempotency mismatch, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
  schemas:
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      additionalProperties: false
      properties:
        refreshToken: { type: string, minLength: 32, maxLength: 256, writeOnly: true }
    AuthResponse:
      type: object
      required: [data, requestId]
      properties:
        data: { type: object, required: [accessToken, expiresAt], properties: { accessToken: { type: string, writeOnly: true }, expiresAt: { type: string, format: date-time } } }
        requestId: { type: string }
    AgentCreate:
      type: object
      required: [name, definition]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        definition: { type: object, additionalProperties: true, maxProperties: 50 }
...
    Agent:
      type: object
      required: [id, tenantId, name, status, version, createdAt, updatedAt]
      properties:
        id: { type: string }
        tenantId: { type: string }
        name: { type: string }
        status: { type: string, enum: [draft, published, archived] }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AgentVersion:
      type: object
      required: [id, agentId, version, contentDigest, publishedAt]
      properties:
        id: { type: string }
        agentId: { type: string }
        version: { type: integer, minimum: 1 }
        contentDigest: { type: string, pattern: '^sha256:[a-f0-9]{64}$' }
        publishedAt: { type: string, format: date-time }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, VERSION_CONFLICT, RATE_LIMIT_EXCEEDED, QUOTA_EXCEEDED, BUDGET_EXCEEDED, INVALID_STATE_TRANSITION, APPROVAL_ALREADY_HANDLED, RESUME_WINDOW_EXPIRED, MODEL_PROVIDER_ERROR, TOOL_EXECUTION_FAILED, SANDBOX_TIMEOUT, INTERNAL_ERROR] }
            message: { type: string, maxLength: 500 }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
    AgentEnvelope: { type: object, required: [data, requestId], properties: { data: { $ref: '#/components/schemas/Agent' }, requestId: { type: string } } }
    AgentVersionEnvelope: { type: object, required: [data, requestId], properties: { data: { $ref: '#/components/schemas/AgentVersion' }, requestId: { type: string } } }
...
    AgentPage: { type: object, required: [data, requestId, meta], properties: { data: { type: array, items: { $ref: '#/components/schemas/Agent' } }, requestId: { type: string }, meta: { $ref: '#/components/schemas/PageMeta' } } }
...
    PageMeta: { type: object, required: [nextCursor], properties: { nextCursor: { type: string, nullable: true } } }
...